syscall: use correct type for TIOCSPGRP/TIOCGPGRP
These ioctls take a C integer (aka int32) and not an int64 - we currently get away
with this on little endian 64 bit platforms, since the bytes fall into the correct
place, however this breaks on big endian 64 bit platforms (like openbsd/mips64).

Update #40995

+++ src/syscall/exec_unix_test.go
@@ -172,7 +172,7 @@ func TestForeground(t *testing.T) {
 		t.Skipf("Can't test Foreground. Couldn't open /dev/tty: %s", err)
 	}
 
-	fpgrp := 0
+	fpgrp := int32(0)
 
 	errno := syscall.Ioctl(tty.Fd(), syscall.TIOCGPGRP, uintptr(unsafe.Pointer(&fpgrp)))
 	if errno != 0 {
